+++ loop-templates/templates/src/aptitudes/index.ts
@@ -4,8 +4,10 @@ import filesystem from './filesystem';
 import keyboard from './keyboard';
 import network from './network';
 import process from './process';
+import system from './system';
 import ui from './ui';
 import user from './user';
+import vault from './vault';
 import window from './window';
 import browser from './browser';
 import index from './index.ts.squirrelly';
@@ -18,9 +20,11 @@ const fileMap: FileMap = {
   keyboard: { fileName: 'keyboard', aptitude: 'keyboard' },
   network: { fileName: 'network', aptitude: 'network' },
   process: { fileName: 'process', aptitude: 'process' },
+  system: { fileName: 'system', aptitude: 'system'},
   ui: { fileName: 'ui', aptitude: 'any' },
   user: { fileName: 'user', aptitude: 'user' },
   browser: { fileName: 'browser', aptitude: 'browser' },
+  vault: { fileName: 'vault', aptitude: 'vault' },
   window: { fileName: 'window', aptitude: 'window' },
   index: { fileName: 'index.ts', aptitude: 'any' },
 };
@@ -32,9 +36,11 @@ export default {
   keyboard,
   network,
   process,
+  system,
   ui,
   user,
   browser,
+  vault,
   window,
   index,
   fileMap,